rm util/index_helpers
