+++ query/src/datasources/table/fuse/index/min_max.rs
@@ -13,6 +13,7 @@
 //  limitations under the License.
 //
 
+use common_base::BlockingWait;
 use common_datavalues::DataSchemaRef;
 use common_exception::ErrorCode;
 use common_planners::Extras;
@@ -98,4 +99,15 @@ impl MinMaxIndex {
 
         Ok(res.into_iter().flatten().collect())
     }
+}
+
+pub fn range_filter(
+    table_snapshot: &TableSnapshot,
+    schema: DataSchemaRef,
+    push_down: Option<Extras>,
+    meta_reader: MetaInfoReader,
+) -> common_exception::Result<Vec<BlockMeta>> {
+    let range_index = MinMaxIndex::new(table_snapshot, &meta_reader);
+    async move { range_index.apply(schema, push_down).await }
+        .wait_in(meta_reader.runtime(), None)?
 }
